% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bib2acad.R
\name{bib2acad}
\alias{bib2acad}
\title{Convert .bib Files to .md Files for the Hugo Academic Theme}
\usage{
bib2acad(bibfile = "", outfold = "", bibfold = "", copybib = FALSE,
  abstract = TRUE, overwrite = FALSE)
}
\arguments{
\item{bibfile}{A string in the format "path/to/bibfile": required}

\item{outfold}{A string in the format "path/to/outfold": required,
the folder must exist}

\item{bibfold}{A string "path/to/bibfold": optional,
if used: the folder must exist}

\item{copybib}{Logical, optional}

\item{abstract}{Logical, optional}

\item{overwrite}{Logical, optional}
}
\description{
\code{bib2acad} takes a .bib file and covert it to a .md file for the
publication folder of the Hugo Academic theme
(\url{https://sourcethemes.com/academic/})
}
